requirements.txt
beautifulsoup4==4.7.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cssselect==1.0.3
dateparser
feedfinder2==0.0.4
feedparser==5.2.1
Flask==1.0.2
Flask-Cors==3.0.7
google-search-results
hammock
idna==2.8
itsdangerous==1.1.0
jieba3k==0.35.1
Jinja2==2.10
jsonlines
lxml==4.6.3
MarkupSafe==1.1.1
marshmallow
newspaper3k==0.2.8
nltk
oauthlib==3.0.1
Pillow==6.2.2
politifact
python-dateutil==2.8.0
pysqlite3
python-dotenv
stanford-corenlp
PyYAML==5.1
requests==2.25.1
requests-file==1.4.3
requests-oauthlib==1.2.0
Scrapy
singledispatch==3.4.0.3
six==1.12.0
soupsieve==1.8
tinysegmenter==0.3
tldextract==2.2.1
tqdm==4.31.1
twython==3.8.0
urllib3==1.26.4
Werkzeug==0.15.1
